requirements.txt
guppy==0.1.10
extruct
graphqlclient
python-dateutil
htmlmin
tomd